arch(policy): split template into files
