+++ api/services/policy/src/engine/templates/index.ts
@@ -1,0 +1,7 @@
+export { eventTrafficLimitPolicy } from './eventTrafficLimitPolicy';
+export { freeTravelForPassengerPolicy } from './freeTravelForPassengerPolicy';
+export { nonFinancialIncentivePolicy } from './nonFinancialIncentivePolicy';
+export { weekdayTrafficLimitPolicy } from './weekdayTrafficLimitPolicy';
+export { financialIncentivePolicy } from './financialIncentivePolicy';
+export { pollutionLimitPolicy } from './pollutionLimitPolicy';
+export { weekendTrafficLimitPolicy } from './weekendTrafficLimitPolicy';
