+++ api/services/policy/src/engine/templates/pollutionLimitPolicy.ts
@@ -1,0 +1,131 @@
+export const pollutionLimitPolicy = {
+  parent_id: null,
+  status: 'template',
+  name: 'Limiter la pollution',
+  slug: 'limiter_la_pollution',
+  description: "Campagne d'incitation financière activable en cas de pic de pollution pour encourager le covoiturage.",
+  global_rules: [
+    {
+      slug: 'weekday_filter',
+      parameters: [0, 1, 2, 3, 4, 5, 6],
+    },
+    {
+      slug: 'rank_whitelist_filter',
+      parameters: ['A', 'B', 'C'],
+    },
+    {
+      slug: 'distance_range_filter',
+      parameters: {
+        min: 2000,
+        max: 150000,
+      },
+    },
+    {
+      slug: 'max_trip_per_target_restriction',
+      parameters: {
+        target: 'driver',
+        amount: 8,
+        period: 'day',
+      },
+    },
+    {
+      slug: 'max_trip_per_target_restriction',
+      parameters: {
+        target: 'passenger',
+        amount: 2,
+        period: 'day',
+      },
+    },
+  ],
+  rules: [
+    [
+      {
+        slug: 'distance_range_filter',
+        parameters: {
+          min: 0,
+          max: 50000,
+        },
+      },
+      {
+        slug: 'passenger_only_filter',
+        parameters: true,
+      },
+      {
+        slug: 'cost_based_amount_setter',
+        parameters: true,
+      },
+    ],
+    [
+      {
+        slug: 'distance_range_filter',
+        parameters: {
+          min: 0,
+          max: 50000,
+        },
+      },
+      {
+        slug: 'driver_only_filter',
+        parameters: true,
+      },
+      {
+        slug: 'fixed_amount_setter',
+        parameters: 10,
+      },
+      {
+        slug: 'per_passenger_modifier',
+        parameters: true,
+      },
+      {
+        slug: 'per_km_modifier',
+        parameters: true,
+      },
+    ],
+    [
+      {
+        slug: 'distance_range_filter',
+        parameters: {
+          min: 50000,
+          max: 150000,
+        },
+      },
+      {
+        slug: 'passenger_only_filter',
+        parameters: true,
+      },
+      {
+        slug: 'fixed_amount_setter',
+        parameters: 500,
+      },
+    ],
+    [
+      {
+        slug: 'distance_range_filter',
+        parameters: {
+          min: 50000,
+          max: 150000,
+        },
+      },
+      {
+        slug: 'driver_only_filter',
+        parameters: true,
+      },
+      {
+        slug: 'fixed_amount_setter',
+        parameters: 500,
+      },
+      {
+        slug: 'per_passenger_modifier',
+        parameters: true,
+      },
+    ],
+  ],
+  ui_status: {
+    for_driver: true,
+    for_passenger: true,
+    for_trip: false,
+    staggered: true,
+  },
+  start_date: null,
+  end_date: null,
+  unit: 'euro',
+};
